Overhaul to GameController, new Node class for DLL, bugfixes for others
Fixed bug where project name was commented out for most classes. Made a Node class for our DoubleLinkedList. Added reverse flag to GameController and a skip flag to the player. Added a PlayFieldView, which peeks the top card of the playField stack. Made our 4 players private fields in GameController to ease initialization of data. Changed how player names are gathered to reflect this.