Changes to texture loading
Texture is now a private field of the card, which is initialized on construction of the card. When a card is created so is it's texture, which is tied to a button when the card is drawn. Implemented a hand sort, but this doesn't work with sorting the observed hand, as I am struggling to sort the textures, for the presentation commenting out the sortHand() call will disable this bug.